% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_processed.R
\name{compute_processed}
\alias{compute_processed}
\title{Compute processed spectral data}
\usage{
compute_processed(spdata)
}
\arguments{
\item{spdata}{data.frame containing the spectral data with the columns
'scope', 'dark', and 'white'}
}
\description{
Conmpute processed spectral data, from the raw count/scope data, counts from
a dark reference, and from a white reference.
}
