% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch_parser.R
\name{dispatch_parser}
\alias{dispatch_parser}
\title{Internal function to dispatch files to the correct parser}
\usage{
dispatch_parser(filename, decimal = ".", sep = NULL)
}
\arguments{
\item{filename}{Path of the file to parse}

\item{decimal}{Character to be used to identify decimal plates
(defaults to \code{.}).}

\item{sep}{Column delimiting characters to be considered in addition to the
default (which are: tab, space, and ";")}
}
\description{
Internal function to dispatch files to the correct parser
}
\keyword{internal}
