% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmetadata.R
\name{getmetadata}
\alias{getmetadata}
\title{Extract metadata from all spectra files in a folder}
\usage{
getmetadata(where = getwd(), ext = "ProcSpec", subdir = FALSE,
  subdir.names = FALSE, cores = getOption("mc.cores", 2L),
  ignore.case = TRUE)
}
\arguments{
\item{where}{(required) folder in which files are located.}

\item{ext}{file extension to be searched for, without the "."
(defaults to "ProcSpec").}

\item{subdir}{should subdirectories within the where folder be included in
the search? (defaults to \code{FALSE}).}

\item{subdir.names}{should subdirectory path be included in the name of the
spectra? (defaults to \code{FALSE}).}

\item{cores}{Number of cores to be used. If greater than 1, import will use
parallel processing (not available in Windows).}

\item{ignore.case}{Logical. Should the extension search be case insensitive?
(defaults to \code{TRUE})}
}
\description{
Extract metadata from all spectra files in a folder
}
\references{
White TE, Dalrymple RL, Noble DWA, O'Hanlon JC, Zurek DB,
Umbers KDL. Reproducible research in the study of biological coloration.
Animal Behaviour. 2015 Aug 1;106:51–7 (\doi{10.1016/j.anbehav.2015.05.007}).
}
