% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightR-package.R
\docType{package}
\name{lightR-package}
\alias{lightR}
\alias{lightR-package}
\title{lightR: Read Spectrometric Data in R}
\description{
Parse various spectra file formats to extract spectral data and metadata, such as integration times, number of averages, spectrometer model or timestamp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bisaloo/lightR}
  \item Report bugs at \url{https://github.com/Bisaloo/lightR/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (0000-0002-4094-1476)

Authors:
\itemize{
  \item Rafael Maia \email{rm72@zips.uakron.edu} (0000-0002-7563-9795)
  \item Thomas White \email{thomas.white026@gmail.com} (0000-0001-8493-9450)
}

}
\keyword{internal}
