% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spec.R
\name{lr_get_spec}
\alias{lr_get_spec}
\title{Extract spectral data from spectra files}
\usage{
lr_get_spec(
  where = getwd(),
  ext = "txt",
  lim = c(300, 700),
  subdir = FALSE,
  subdir.names = FALSE,
  ignore.case = TRUE,
  interpolate = TRUE,
  parser = NULL,
  ...
)
}
\arguments{
\item{where}{Folder in which files are located (defaults to current working
directory).}

\item{ext}{File extension to be searched for, without the "." (defaults to
\code{txt}). You can also use a character vector to specify multiple file
extensions.}

\item{lim}{A vector with two numbers determining the wavelength limits to be
considered (defaults to \code{c(300, 700)}).}

\item{subdir}{Should subdirectories within the \code{where} folder be included in
the search? (defaults to \code{FALSE}).}

\item{subdir.names}{Should subdirectory path be included in the name of the
spectra? (defaults to \code{FALSE}).}

\item{ignore.case}{Should the extension search be case insensitive? (defaults
to \code{TRUE})}

\item{interpolate}{Boolean indicated whether spectral data should be
interpolated and pruned at every nanometre. Note that this option can only
work if all input data samples the same wavelengths. Defaults to \code{TRUE}.}

\item{parser}{Optional function to specify the parser to be used. By default,
the parser is automatically selected based on the file extension. This argument
can be used to force the use of a specific parser, or to pass a custom parser.
See details for more information on custom parsers. Please note that this
argument can produce easily produce invalid results if used incorrectly and
only advanced users should use it, with extra caution.}

\item{...}{Arguments passed to individual parsers.}
}
\value{
A data.frame, containing the wavelengths in the first column and
individual imported spectral files in the subsequent columns.
}
\description{
Finds and imports reflectance/transmittance/absorbance data from spectra
files in a given location.
}
\details{
\subsection{Parallel processing}{

You can customise the type of parallel processing used by this function with
the \code{\link[future:plan]{future::plan()}} function. This works on all operating systems, as well
as high performance computing (HPC) environment. Similarly, you can customise
the way progress is shown with the \code{\link[progressr:handlers]{progressr::handlers()}} functions
(progress bar, acoustic feedback, nothing, etc.)
}

\subsection{Custom parsers}{

To create a custom parser to pass to the \code{parser} argument, you need to create
with the following signature:
\itemize{
\item input: a file path (string) as a first argument, and optionally, any
additional arguments needed to parse the file. These arguments can be passed
to the \code{lr_get_spec()} function via the \code{...} argument.
\item output: a named list of two elements, as defined in \code{lr_parse_generic()}.
}
}
}
\examples{
spcs <- lr_get_spec(system.file("testdata", package = "lightr"), ext = "jdx")
head(spcs)

}
\seealso{
\code{\link[pavo:getspec]{pavo::getspec()}}
}
