% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_jdx.R
\name{lr_parse_jdx}
\alias{lr_parse_jdx}
\title{Parse OceanOptics JCAMP-DX (.jdx) file}
\usage{
lr_parse_jdx(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse OceanOptics JCAMP-DX (.jdx) file. \url{https://oceanoptics.com/}
}
\examples{
lr_parse_jdx(system.file("testdata", "OceanOptics.jdx", package = "lightr"))

}
\references{
McDonald RS, Wilks PA. JCAMP-DX: A Standard Form for Exchange of
Infrared Spectra in Computer Readable Form. Applied Spectroscopy.
1988;42(1):151‑62.
}
