% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_procspec.R
\name{lr_parse_procspec}
\alias{lr_parse_procspec}
\title{Parse OceanInsight ProcSpec file}
\usage{
lr_parse_procspec(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse OceanInsight (formerly OceanOptics) ProcSpec file.
\url{https://www.oceaninsight.com/}
}
\examples{
lr_parse_procspec(system.file("testdata", "procspec_files",
                              "OceanOptics_Linux.ProcSpec",
                              package = "lightr"))

}
\references{
\url{https://www.oceaninsight.com/support/faqs/software/}
}
