% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_craic_binary.R
\name{lr_parse_spc}
\alias{lr_parse_spc}
\title{Parse CRAIC binary file}
\usage{
lr_parse_spc(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse CRAIC binary file. \url{http://www.microspectra.com}
}
\examples{
lr_parse_spc(system.file("testdata", "compare", "CRAIC", "CRAIC.spc",
                         package = "lightr"))

}
