% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avantes_binary.R
\name{lr_parse_trm}
\alias{lr_parse_trm}
\alias{lr_parse_abs}
\alias{lr_parse_roh}
\title{Parse Avantes binary file}
\usage{
lr_parse_trm(filename)

lr_parse_abs(filename)

lr_parse_roh(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse Avantes binary file. \url{https://www.avantes.com/products/spectrometers}
}
\examples{
lr_parse_trm(system.file("testdata", "avantes_trans.TRM",
                         package = "lightr"))
lr_parse_roh(system.file("testdata", "avantes_reflect.ROH",
                         package = "lightr"))

}
