% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avantes_converted.R
\name{lr_parse_ttt}
\alias{lr_parse_ttt}
\alias{lr_parse_trt}
\title{Parse Avantes converted file}
\usage{
lr_parse_ttt(filename)

lr_parse_trt(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse Avantes converted file.
\url{https://www.avantes.com/products/spectrometers}
}
\examples{
lr_parse_ttt(system.file("testdata", "avantes_export.ttt",
                         package = "lightr"))
lr_parse_trt(system.file("testdata", "avantes_export2.trt",
                         package = "lightr"))

}
