% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_generic.R
\name{parse_generic}
\alias{parse_generic}
\title{Generic function to parse spectra files that don't have a specific parser}
\usage{
parse_generic(filename, decimal, sep)
}
\arguments{
\item{filename}{Path of the file to parse}

\item{decimal}{Decimal separator}

\item{sep}{Column separator}
}
\value{
A list of two elements:
  * a dataframe with columns "wl", "dark", "white", "scope" in that order
  * a list with metadata including
}
\description{
Generic function to parse spectra files that don't have a specific parser
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}

Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
