% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_oceanoptics_converted.R
\name{parse_jaz}
\alias{parse_jaz}
\alias{parse_jazirrad}
\title{Parse OceanOptics converted file}
\usage{
parse_jaz(filename)

parse_jazirrad(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" in that order
\item a list with metadata including
}
}
\description{
Parse OceanOptics converted file
}
\examples{
parse_jaz(system.file("testdata", "jazspec.jaz", package = "lightr"))
parse_jazirrad(system.file("testdata", "irrad.JazIrrad", package = "lightr"))

}
