% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_jdx.R
\name{parse_jdx}
\alias{parse_jdx}
\title{Parse OceanOptics JCAMP-DX (.jdx) file}
\usage{
parse_jdx(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
  * a dataframe with columns "wl", "dark", "white", "scope" in that order
  * a list with metadata including
}
\description{
Parse OceanOptics JCAMP-DX (.jdx) file
}
\references{
McDonald RS, Wilks PA. JCAMP-DX: A Standard Form for Exchange of
Infrared Spectra in Computer Readable Form. Applied Spectroscopy.
1988;42(1):151‑62.
}
