% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_oceanoptics_converted.R
\name{parse_oceanoptics}
\alias{parse_oceanoptics}
\alias{parse_jaz}
\alias{parse_jazirrad}
\title{Parse OceanOptics converted file}
\usage{
parse_oceanoptics(filename)

parse_jaz(filename)

parse_jazirrad(filename)
}
\arguments{
\item{filename}{Path of the ABS, TRM or ROH file}
}
\value{
A list of two elements:
  * a dataframe with columns "wl", "dark", "white", "scope" in that order
  * a list with metadata including
}
\description{
Parse OceanOptics converted file
}
\author{
Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
