% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avantes_binary.R
\name{parse_trm}
\alias{parse_trm}
\alias{parse_abs}
\alias{parse_roh}
\title{Parse Avantes binary file}
\usage{
parse_trm(filename)

parse_abs(filename)

parse_roh(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" in that order
\item a list with metadata including
}
}
\description{
Parse Avantes binary file
}
