% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avantes_converted.R
\name{parse_ttt}
\alias{parse_ttt}
\alias{parse_trt}
\title{Parse Avantes converted file}
\usage{
parse_ttt(filename)

parse_trt(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
  * a dataframe with columns "wl", "dark", "white", "scope" in that order
  * a list with metadata including
}
\description{
Parse Avantes converted file
}
