% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2csv.R
\name{spec2csv}
\alias{spec2csv}
\title{Convert spectral data files to csv files}
\usage{
spec2csv(filename, overwrite = FALSE)
}
\description{
Convert spectral data files to csv files
}
\section{Warning}{


This step loses all metadata associated to the spectra. This metadata is
critical to ensure reproducibility. We recommended you use \code{\link[=getmetadata]{getmetadata()}} to
extract this information from your raw data
}

