% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightr-package.R
\docType{package}
\name{lightr-package}
\alias{lightr}
\alias{lightr-package}
\title{lightr: Read Spectrometric Data and Metadata}
\description{
Parse various reflectance/transmittance/absorbance spectra file
    formats to extract spectral data and metadata, as described in Gruson, White
    & Maia (2019) <doi:10.21105/joss.01857>.
    This package has been peer-reviewed by rOpenSci (v. 0.1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/lightr}
  \item \url{https://github.com/ropensci/lightr}
  \item Report bugs at \url{https://github.com/ropensci/lightr/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

Authors:
\itemize{
  \item Rafael Maia (\href{https://orcid.org/0000-0002-7563-9795}{ORCID})
  \item Thomas White (\href{https://orcid.org/0000-0002-3976-1734}{ORCID})
}

Other contributors:
\itemize{
  \item Kotya Karapetyan (Author of the MATLAB script to read Avantes binary files (CC-BY)) [copyright holder]
}

}
\keyword{internal}
