% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightr-package.R
\docType{package}
\name{lightr-package}
\alias{lightr}
\alias{lightr-package}
\title{lightr: Read Spectrometric Data in R}
\description{
Parse various UV-VIS reflectance/transmittance/absorbance spectra
    file formats to extract spectral data and metadata.
    This package has been peer-reviewed by rOpenSci (v. 0.1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/lightr}
  \item \url{https://github.com/ropensci/lightr}
  \item Report bugs at \url{https://github.com/ropensci/lightr/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (0000-0002-4094-1476)

Authors:
\itemize{
  \item Rafael Maia (0000-0002-7563-9795)
  \item Thomas White (0000-0001-8493-9450)
}

}
\keyword{internal}
