% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_oceanoptics_converted.R
\name{lr_parse_jaz}
\alias{lr_parse_jaz}
\alias{lr_parse_jazirrad}
\title{Parse OceanInsight converted file}
\usage{
lr_parse_jaz(filename)

lr_parse_jazirrad(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse OceanInsight (formerly OceanOptics) converted file.
\url{https://www.oceaninsight.com/}
}
\examples{
lr_parse_jaz(system.file("testdata", "jazspec.jaz", package = "lightr"))
lr_parse_jazirrad(system.file("testdata", "irrad.JazIrrad",
                  package = "lightr"))

}
