% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_oceanoptics_converted.R
\name{parse_oceanoptics}
\alias{parse_oceanoptics}
\alias{parse_jaz}
\alias{parse_jazirrad}
\title{Parse OceanOptics converted file}
\usage{
parse_oceanoptics(filename)

parse_jaz(filename)

parse_jazirrad(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\description{
Parse OceanOptics converted file
}
