% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_procspec.R
\name{parse_procspec}
\alias{parse_procspec}
\title{Import ProcSpec spectra file}
\usage{
parse_procspec(filename)
}
\arguments{
\item{filename}{Path of the ProcSpec file}
}
\value{
A list of two elements:
  * a dataframe with columns "wl", "dark", "white", "scope" in that order
  * a list with metadata including
}
\description{
Import ProcSpec spectra file
}
\references{
https://oceanoptics.com/faq/extract-data-procspec-file-without-spectrasuite/
}
\author{
Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
