% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_procspec.R
\name{parse_procspec}
\alias{parse_procspec}
\title{Parse OceanOptics ProcSpec file}
\usage{
parse_procspec(filename)
}
\arguments{
\item{filename}{Path of the file to parse}
}
\value{
A list of two elements:
\itemize{
\item a dataframe with columns "wl", "dark", "white", "scope" and "processed",
in that order
\item a list with metadata including
}
}
\description{
Parse OceanOptics ProcSpec file. \url{https://oceanoptics.com/}
}
\examples{
parse_procspec(system.file("testdata", "procspec_files",
                           "OceanOptics_Linux.ProcSpec", package = "lightr"))

}
\references{
\url{https://oceanoptics.com/faq/extract-data-procspec-file-without-spectrasuite/}
}
